feat(benchmark): cleanup
